Bug 1657690 - Fix an edge case backface visibility regression. r=kvark
Ensure the current ancestor spatial node for backface visibility
calculations is tracked, even when an otherwise redundant stacking
context is removed.

The code handling flattened stacking contexts is overcomplicated,
due to limitations in the current public API. For now, this seems
like a reasonable fix, although long term we should adapt the
public API so that the internals are not so complex (specifically,
the current stacking context API serves many related purposes, and
should be split up into more explicit APIs for defining a 3d context,
filter list etc).

Differential Revision: https://phabricator.services.mozilla.com/D86499