Bug 1664719 - Pt 1 - Select outermost scroll root if no real scroll roots found. r=nical
Previously, if we encounter a page where there are no real scroll
roots, WR would attach the picture cache slices to the root reference
frame.

This means that in some cases, no shared clips will be found for the
content slices. This is not currently a performance issue, since the
content slices are fixed and don't scroll. However, future patches
for this bug will only select shared clips that are ancestors of
the picture cache scroll root.

This patch ensures that in these cases, we still select the outermost
defined scroll frame, allowing shared clips to be correctly found
for these pages with a single scroll frame that doesn't scroll.

Differential Revision: https://phabricator.services.mozilla.com/D90051