Bug 1668472 - Ensure non-presented frames don't interfere with partial present logic. r=gw
Sometimes webrender will render a frame that will not be presented,
such as when a document is being replaced but its frame must be
rendered to update the texture cache. This interferes with the simple
compositor's partial present's damage region tracking logic, as it
needs to track damage between presents, but instead tracks damage
between renders.

To fix this, ensure that we no longer composite such frames to the
main framebuffer, only render their offscreen targets. This was
historically the case, but changed when we briefly rendered picture
cache tiles directly to the main framebuffer then blitted them to the
texture cache. We no longer do this, so this can be safely reverted.

Rather than trying to track picture cache tile dirty regions across
multiple renders, simply force a full present on the next frame
whenever we encounter a frame that will itself not be presented.

Differential Revision: https://phabricator.services.mozilla.com/D92122