Bug 1682365 - Pt 6 - Add new frame graph implementation. r=nical
This patch introduces the new frame graph implementation, which
allows for more advanced and efficient render task graphs.

The goal of this initial work is to achieve feature parity with
the existing render task graph.

Follow up work will take advantage of the new graph functionality
to improve the efficiency of current mix-blend-mode, backdrop-filter
and svg filter operations.

Differential Revision: https://phabricator.services.mozilla.com/D99743