Bug 1741779 - Temporarily make backdrop-filter display item a no-op r=gfx-reviewers,nical
It's disabled by preference due to some bugs that require some
significant refactoring work.

The current implementation breaks a couple of assumptions we want
to make in WR related to clip vs. prim hierarchies (and is the only
primitive which does this). Disabling backdrop-filter in the interim
allows us to refactor these parts of WR, and then update the filter
implementation and re-enable it afterwards.

Differential Revision: https://phabricator.services.mozilla.com/D131462