Bug 1743152 - Handle duplicate iframe references in display lists r=gfx-reviewers,nical
In certain situations (such as a print preview dialog, with fission
enabled, where an iframe spans a page), the display list contains
a reference to the same iframe multiple times.

Handle this in WR by namespacing the spatial id (via a stack as
iframes are traversed) and uid (by iframe pipeline instance id).

This is more of a band-aid fix than a proper fix - we need to
rethink how we handle spatial mapping properly with iframe
instances.

Differential Revision: https://phabricator.services.mozilla.com/D132661