Bug 1748499 - Switch WR's scale_factors to operate in double precision r=nical
This function computes squares of values, but the results are the same magnitude than the source components.
In fact, they can't be larger than the biggest component. Therefore, by doubling the precision of computation we can guarantee that the result will be finite if the source data is finite.

An alternative approach would be to figure out a different computation path that would preserve the precision as it goes. Looks to not be necessary here.

Differential Revision: https://phabricator.services.mozilla.com/D135140