Bug 1834678 - Upgrade meson to 1.1.0. r=gfx-reviewers,gw
The old version that is currently used doesn't support lld for mac, and
tries to pass it flags for BFD ld (e.g. --as-needed).

For some reason meson checks fail because of some CFLAGS not being used
during the checks being reported as an error, so we make those errors
quiet.

Differential Revision: https://phabricator.services.mozilla.com/D178899