Bug 1849530 - Add support for sub-passes on a single render task r=gfx-reviewers,ahale
This adds support for sub-passes to individual render tasks, which
allow subsequent rendering to be configured on top of an existing
render task.

This patch introduces no functional change, it's preparation for
follow up patches.

The follow up patches will make use of this to apply clip masks to
picture primitives directly, where applicable, rather than allocating
a screen-space clip mask and drawing to that.

Differential Revision: https://phabricator.services.mozilla.com/D186614