Bug 1859100 - Correctly track whether any ancestor spatial node is zooming. r=gw
When updating the SpatialTree we record whether each spatial node is
itself or any of its ancestors are actively async zooming. This allows
us to make certain optimizations, such as low-quality pinch zoom mode,
or rounding the raster scale for glyphs.

The current implementation of this, however, contains a bug which
means that a node's is_ancestor_or_self_zooming variable will only be
true when itself or its _direct_ ancestor are zooming, rather than
_any_ ancestor. On certain pages this means we are unable to make the
aforementioned optimizations, resulting in poor performance. This
patch rectifies this problem and adds a unit test.

Differential Revision: https://phabricator.services.mozilla.com/D203446