Bug 2001243 - Remove usage of deprecated dwrote functions in wr_glyph_rasterizer. r=gfx-reviewers,lsalzman
Since we updated the dwrote crate in bug 2000190, we can switch to newer versions
of these functions that return failures to the caller instead of just panic!()ing
in the event of a DirectWrite error.

I'm not sure how well we handle such failures -- we'll probably end up rendering
garbled text in cases where an expected font fails to load -- but that's preferable
to completely crashing the GPU process.

Differential Revision: https://phabricator.services.mozilla.com/D276688